requirements.txt
amqp==5.0.9
billiard==3.6.4.0
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
google-api-core==1.21.0
google-auth==1.18.0
google-cloud-core==1.3.0
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
greenlet==1.0.0
kombu==5.2.3
pika==1.1.0
playwright==1.17.2
prompt-toolkit==3.0.24
pyee==8.2.2
pytz==2021.3
six==1.16.0
typing-extensions==4.0.1
vine==5.0.0
wcwidth==0.2.5
websockets==10.1
